Разделила код на модули и создала файлы с возможными будущими модулями
