+++ js/form.js
@@ -1,0 +1,1 @@
+// Заполнение формы пожачи объявления, валидация и отправка формы на сервер
