+++ js/map.js
@@ -1,0 +1,1 @@
+//описывает поведение метки
