+++ js/page.js
@@ -1,0 +1,1 @@
+//Активное/неактивное состояние элементов
